`sappho-east`: split out common expr types into individual mods.
